+++ NEWS.md
@@ -1,3 +1,5 @@
 # demspaces 0.0.0.9000
 
 * Added a `NEWS.md` file to track changes to the package.
+* Added a normalization option for `ds_logistic_reg()`. This uses `make_standardizer()` to generate a function that will standardize any non-binary predictor columns, using the training data means and SDs. 
+